add filter and remove logs
